COSSIM
COSSIM Simulation Framework is an open-source framework that efficiently integrate a series of sub-tools that model the computing devices of the processing nodes as well as the network(s) of the parallel systems. It provides cycle accurate results by simulating the actual application and system software executed on each node, realistic communication modeling of the different networks that are employed.